[Refactor] 중고 관련 response 네이밍 변경 #183
